adding torsion cluster tests, split up testing scripts
